temp move shareHandler function to DiscussionProvider
